Create a proof of concept for playing back rtp from ui_call_raw.
When a SIP message with an SDP header is detected, the ui will show a
button that uses SoX to playback the first detected RTP stream to the
user.

Since sngrep only seems to pass through the actual RTP payload to the
pcap for effiency, opposed to storing them in memory, right now all raw
RTP streams are written to /tmp/sngrep-<callid>-XXXXXX to facilitate
playback at a later time.

Ideally this should be improved such that complete RTP streams will be
retrieved from the pcap file itself.

As a proof of concept this is working quite well, which is why I'd like
to ask maintainers with more extensive experience in C (and sngrep) for
feedback and, if you think this is a great idea, suggestions as to what
approach would work best for retrieving RTP streams on (very) large
scale.

Suggestions regarding how this could be integrated into the ui better
are very welcome as well, and given enough time I might be ambitious
enough to try to write an RTP statistics page a-la Wireshark :)

Regards (and thanks in advance)